Ladder graph logging (#105)
* More detailed log of descartes ladder graph creation

* Add format() and operator<< methods for waypoint sampler

This makes it easier to get info on which sampler failed

* Better failed vertex output

* Output one sampled position when an edge fails

Messy but code completion wouldn't work without being extra redundant in declarations.  Outputting 1 state is enough to pose the robot and see where it was trying to go.  It would be better to pull the original cartesian target pose out but that would be more difficult.

* Changed printing to use ostream across Descartes logging

* Fixed incorrect name of source file

* Clang formatting

* Fix new clang tidy warning

* Fix bgl clang tidy warning

* Fixed an accidental bug because I wasn't building with debug before

* Updated printout to include a std::end so it actually prints

* Revert unneeded clang tidy fix

---------

Co-authored-by: Douglas Smith <douglas.smith@swri.org>